requirements.txt
albumentations
diffusers
dotenv
einops
flask
Flask-Cors
gradio
jupyter
lightning
transformers
torchao
torchvision
openai
opencv-python
peft
prodigyopt
protobuf
requests
scikit-learn
sentencepiece
streamlit
supervision
wandb
git+https://github.com/PramaLLC/BEN2.git